requirements.txt
functions-framework
cloudevents
google-cloud-bigquery
pandas
google-cloud-storage